requirements.txt
# Dipendenze principali
python-telegram-bot==22.0
pandas>=2.0.0
flask>=2.0.0
flask-login>=0.6.0
flask-wtf>=1.0.0
werkzeug==2.3.7
python-dotenv>=0.19.0
supabase>=2.0.0
httpx>=0.26.0

# Dipendenze per la generazione di report
xlsxwriter>=3.0.0
openpyxl>=3.1.0
reportlab>=4.0.0

# Dipendenze per il deployment
gunicorn>=20.0.0
aws-lambda-wsgi>=0.0.6
